move plexscoring -> search
