+++ tests/unit_tests/search/scoring/test_features.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from alphadia.plexscoring.features.fragment_features import center_envelope_1d
+from alphadia.search.plexscoring.features.fragment_features import center_envelope_1d
 
 
 @pytest.mark.parametrize(
